requirements.txt
Django==4.1
Djangorestframework==3.13.0
Pandas==1.4.4
psycopg2==2.9.3
ipython==8.4.0